# docs/swagger.yaml
openapi: 3.0.2
info:
  title: Barbank
  version: 1.0.0
paths:
...
  /transactions:
    summary: Path used to manage the list of transactions.
    description: >-
      The REST endpoint/path used to list and create zero or more `transaction` entities.  This path
      contains a `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `transaction` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/transaction'
        required: true
      tags:
        - transactions
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid amount:
                  value:
                    error: Invalid amount
                Missind parameter:
                  value:
                    error: 'Transaction validation failed: .* : Path `.*` is required.'
                'Path `accounFrom`  (`a`) is shorter than the minimum allowed length. ':
                  value:
                    error: 'Path `.*`  (`.*`) is shorter than the minimum allowed length. '
                'Transaction validation failed: explanation: Path `explanation` is required.':
                  value:
                    error: >-
                      Transaction validation failed: explanation: Path `explanation` is
                      required.
          description: Bad request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid Authorization header format:
                  value:
                    error: Invalid Authorization header
                Missing Authorization header:
                  value:
                    error: Missing Authorization header
          description: Problem with authorization header
        '402':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Insufficient funds:
                  value:
                    error: Insufficient funds
          description: Insufficient funds
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                accountFrom not found:
                  value:
                    error: accountFrom not found
          description: Not found
      security:
        -
          Bearer: []
      operationId: createtransaction
      description: Creates a new instance of a transaction
  /transactions/b2b:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/jwt'
        required: true
      tags:
        - transactions
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/transaction'
              examples:
                POST /transactions/b2b response:
                  value:
                    receiverName: Taivo
          description: Transaction accepted
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid signature:
                  value:
                    error: Invalid signature
          description: Invalid signature
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Bank \w+ was not found in Central Bank:
                  value:
                    error: Bank \w+ was not found in Central Bank
          description: Bank was not found in Central Bank
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                There was an error communication with Central Bank:
                  value:
                    error: There was an error communication with Central Bank.*
          description: There was an error communication with Central Bank
      summary: Accepts transactions from other banks
...
components:
  schemas:
...
    error:
      title: Root Type for error
      description: ''
      type: object
      properties:
        error:
          type: string
      example:
        error: Error message.
...
    transaction:
      title: Root Type for transaction
      description: ''
      type: object
      properties:
        accountFrom:
          type: string
        accountTo:
          type: string
        amount:
          format: int32
          type: integer
        explanation:
          type: string
      example:
        accountFrom: cdd3207da5457d94ed04c0b5b56cf2b5a33
        accountTo: cddaf622492dd0234b6928678aa1fe0e294
        amount: 0.01
        explanation: Don’t go spend it all at once
    jwt:
      title: Root Type for jwt
      description: ''
      type: object
      properties:
        jwt:
          type: string
      example:
        jwt: >-
          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NvdW50RnJvbSI6ImNkZDMyMDdkYTU0NTdkOTRlZDA0YzBiNWI1NmNmMmI1YTMzIiwiYWNjb3VudFRvIjoiY2RkYWY2MjI0OTJkZDAyMzRiNjkyODY3OGFhMWZlMGUyOTQiLCJhbW91bnQiOjEsImN1cnJlbmN5IjoiVVNEIiwiZXhwbGFuYXRpb24iOiJ4eHgiLCJzZW5kZXJOYW1lIjoiVGFpdm8gTnVyayJ9.cZALqpZE3_fDH0Ofymwzw5LB5gIwPcdUxtm518OLvFY
...
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
...